% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsamples.R
\name{getsamples}
\alias{getsamples}
\title{Extract samples of effect size measures from primary studies}
\usage{
getsamples(
  obj,
  param = NULL,
  package = NULL,
  n.samples = 10000,
  sm,
  transf = TRUE
)
}
\arguments{
\item{obj}{Meta-analysis object (see Details).}

\item{param}{Parameters for effect sizes (see Details).}

\item{package}{Character string with name of R package used to
create the meta-analysis object \code{obj}.}

\item{n.samples}{Number of samples per study if meta-analysis had been
performed with R package \bold{meta} or
\code{\link[metafor]{rma}} from R package \bold{metafor}.}

\item{sm}{A character string indicating the summary measure used in
primary studies.}

\item{transf}{A logical indicating whether effect sizes are
transformed or on the original scale. If \code{transf = TRUE}
(default), effect estimates are expected to be on the log scale
(for example, log odds ratios instead of odds ratios for \code{sm
= "OR"}).}
}
\value{
A matrix containing sampled effect sizes for primary study.
}
\description{
Obtains a matrix containing sampled effect sizes from the the
primary studies.
}
\details{
This function can be used to create a matrix with sampled effect sizes of
primary studies (number of rows equal to the number of samples,
number of columns equal to the number of studies in the meta-analysis).

Main input to this function is argument \code{obj} containing a
meta-analysis object created with \code{bugs} from R package
\bold{R2OpenBUGS}, \code{jags} from \bold{R2jags},
\code{jags.samples} from \bold{rjags}, \code{brm} from \bold{brms},
\code{rma} from \bold{metafor}, or \code{metabin}, \code{metacont}
or\code{metagen} from \bold{meta}.

Argument \code{param} corresponds to the effect size measures of
the primary studies (R2OpenBUGS, R2jags, rjags packages) or to the
name of the column that contains the unique identification of each
primary study in the data frame containing the results of primary
studies used for meta-analysis (\bold{brms} package). This argument
is not required if meta-analysis has been performed using
\bold{meta} or \bold{metafor} package.

Information of the R package used to conduct the meta-analysis can
be provided in argument \code{package}. The value of \code{package}
can be "R2OpenBUGS", "R2jags", "rjags", "meta", "metafor" or
"brms". This information is not required for an object created with
\bold{meta}, \bold{metafor} or \bold{brms}.

Argument \code{n.samples} contains the number of samples per study for
meta-analyses created with \bold{meta} or \bold{metafor}. Default
is 10000 samples.

Summary measure used in primary studies (argument \code{sm}) can be
either "OR" (odds ratio), "RR" (risk ratio), "HR" (hazard ratio),
"RD" (risk difference), "MD" (mean difference), "SMD" (standardised
mean difference), "GEN_ratio" (generic ratio) or "GEN_diff"
(generic difference). Does not need to be provided if meta-analysis
was performed with \bold{meta} or \bold{metafor}.
}
\examples{
\dontrun{
# Example with an object obtained using the R2OpenBUGS package ("es",
# whose simulations on the effect sizes of primary studies are those
# of the "delta" parameter). If the object had been obtained using a
# different package, all remaining arguments would be the same and only
# the "package" argument would have a different input.

# In this example, argument \code{transf = TRUE} (default), as sampled
# effect sizes are log odds ratios.

load(url("https://raw.github.com/BernardoSousaPinto/metainc_extra_files/main/es.Rdata"))

sample <- getsamples(es, param = "delta", package = "R2OpenBUGS", sm = "OR")
sample
}

# Example with a dataset containing variables indicating the effect sizes
# for each primary study (yi) and the respective variances (vi), and for
# which frequentist meta-analysis using the meta package is applied.

data("anticoagulation_df")
m1 <- meta::metagen(yi, sqrt(vi), sm = "OR", data = anticoagulation_df,
  studlab = LETTERS[1:18])
set.seed(1090) # Make sampled effect sizes reproducible
sample1 <- getsamples(m1)
sample1

\dontrun{
# Same samples of effect sizes using R package metafor (must be installed)

m2 <- metafor::rma(anticoagulation_df, measure = "OR", slab = LETTERS[1:18])
set.seed(1090) # Make sampled effect sizes reproducible
sample2 <- getsamples(m2)
sample2

all.equal(sample1, sample2) # Only difference: package name
}

}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt}
}
