% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsims.R
\name{getsims}
\alias{getsims}
\title{Extract simulation results with effect size measures of primary studies}
\usage{
getsims(obj, param = NULL, package = NULL, nsim = 10000, sm, transf = TRUE)
}
\arguments{
\item{obj}{Meta-analysis object (see Details).}

\item{param}{Parameters for effect sizes (see Details).}

\item{package}{Character string with name of R package used to create
meta-analysis object \code{obj}.}

\item{nsim}{Number of simulation runs if meta-analysis had been performed
with \code{\link[metafor]{rma}} from R package \bold{metafor}.}

\item{sm}{A character string indicating the summary measure used in
primary studies.}

\item{transf}{A logical indicating whether effect sizes are transformed or
on the original scale. If \code{transf = TRUE} (default), effect estimates
are expected to be log odds ratios instead of odds ratios for
\code{sm = "OR"}, for example.}
}
\value{
A matrix containing simulated effect sizes for primary study.
}
\description{
Obtains a matrix containing simulation results of effect size measures of
the primary studies.
}
\details{
This function can be used to create a matrix with simulated effect sizes of
primary studies (number of rows equal to the number of simulation runs,
number of columns equal to the number of studies in the meta-analysis.

Main input to this function is a meta-analysis object created with
\code{bugs} from R package \bold{R2OpenBUGS}, \code{jags} from \bold{R2jags},
\code{jags.samples} from \bold{rjags}, \code{brm} from \bold{brms}, or
\code{rma} from \bold{metafor}.

Parameter corresponding to the effect size measures of the primary studies
(R2OpenBUGS, R2jags, rjags packages) or to the name of the column that
contains the unique identification of each primary study in the
data frame containing the results of primary studies used for meta-analysis
(brms package). This argument is not defined if meta-analysis has been
performed using \bold{metafor} package.

Information of the R package used to conduct the meta-analysis can be
provided in argument \code{package}. This information must not be provided
for an object created with \bold{metafor} or \bold{brms}. The value of
\code{package} should be "R2OpenBUGS", "R2jags", or "rjags".

The number of simulation runs (argument \code{nsim}) can be specified for
meta-analysis objects created with \bold{metafor}. Default is 10000 runs.
}
\examples{
\dontrun{
# Example with an object obtained using the R2OpenBUGS package ("es",
# whose simulations on the effect sizes of primary studies are those
# of the "delta" parameter). If the object had been obtained using a
# different package, all remaining arguments would be the same and only
# the "package" argument would have a different input.

# In this example, argument \code{transf = TRUE} (default), as simulation
# results had been obtained as ln(odds ratio).

load(url("https://raw.github.com/BernardoSousaPinto/metainc_extra_files/main/es.Rdata"))

sims <- getsims(es, param = "delta", package = "R2OpenBUGS", sm = "OR")
sims
}

# Example with a dataset containing variables indicating the effect sizes
# for each primary study (yi) and the respective variances (vi), and for
# which frequentist meta-analysis using the meta package is applied.

data("anticoagulation_df")
m1 <- meta::metagen(yi, sqrt(vi), sm = "OR", data = anticoagulation_df,
  studlab = LETTERS[1:18])
set.seed(1090) # Make simulated effect sizes reproducible
sims1 <- getsims(m1)
sims1

\dontrun{
# Same simulated effect sizes using R package metafor (must be installed)

m1f <- metafor::rma(anticoagulation_df, measure = "OR", slab = LETTERS[1:18])
set.seed(1090) # Make simulated effect sizes reproducible
sims2 <- getsims(m1f)
sims2

all.equal(sims1, sims2) # Only difference: package name
}

}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt}
}
