% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grade}
\alias{grade}
\title{Title ...}
\usage{
grade(
  yi,
  vi,
  rob = NULL,
  data = NULL,
  br = NA,
  sm = "GEN",
  scale = 1000,
  threshold.pval.Q = 0.1,
  threshold.I2 = 0.6,
  nsim = 10000,
  keep = TRUE,
  ...
)
}
\arguments{
\item{yi}{...}

\item{vi}{...}

\item{rob}{...}

\item{data}{...}

\item{br}{...}

\item{sm}{...}

\item{scale}{...}

\item{threshold.pval.Q}{...}

\item{threshold.I2}{...}

\item{nsim}{...}

\item{keep}{...}

\item{\dots}{Additional arguments (ignored).}
}
\value{
The \code{grade} function returns a list with
\item{ASI}{...}
\item{DI}{...}
\item{inc.Q}{...}
\item{inc.I2}{...}
\item{pval.Q}{...}
\item{threshold.pval.Q}{...}
\item{I2 = I2}{...}
\item{threshold.I2}{...}
\item{yi}{...}
\item{vi}{...}
\item{rob}{...}
\item{sm}{...}
\item{ES}{...}
\item{intervention.events}{...}
\item{baseline.events}{...}
\item{br}{...}
\item{scale}{...}
}
\description{
One (or few) sentence(s) description.
}
\details{
Details ...
}
\examples{
data(anticoagulation)
inc_anticoagulation <-
  inc(log(anticoagulation), t1 = 20, t2 = 30, t3 = 40, br = 0.5, sm = "or")
sens <- inf(inc_anticoagulation)
sens

}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
