\name{inc-class}


\alias{inc-class}
\alias{summary.inc-class}
\alias{print.inc}
\alias{summary.inc}

\title{
Methods for the class \code{inc}.
}
\description{
Various methods for the output of the function \code{\link{inc}}.
}
\usage{
\method{print}{inc}(x,...)
\method{summary}{inc}(object,...)
}
\arguments{
    \item{x}{An object of class \code{inc}.}
    \item{object}{An object of class \code{inc}.}
    \item{\dots}{arguments to be passed}
}

\seealso{
\code{\link{inc}}
}
