% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_bin.R
\name{inc_bin}
\alias{inc_bin}
\alias{print.inc_bin}
\alias{summary.inc_bin}
\alias{print.summary.inc_bin}
\title{Title line}
\usage{
inc_bin(sims, baseline.risk, t, t1, t2, t3, sm, scale = 1000)

\method{print}{inc_bin}(x, ...)

\method{summary}{inc_bin}(object, ...)

\method{print}{summary.inc_bin}(x, ...)
}
\arguments{
\item{sims}{R object ...}

\item{baseline.risk}{...}

\item{t}{...}

\item{t1}{...}

\item{t2}{...}

\item{t3}{...}

\item{sm}{... "or", "rr", or "hr"  ...}

\item{scale}{...}

\item{x}{R object of class \code{inc_bin} or  \code{summary.inc_bin}.}

\item{\dots}{Additional arguments (ignored)}

\item{object}{R object of class \code{inc_bin}.}
}
\value{
Description of returned R object (data frame?)
}
\description{
Brief description (more details could be provided under Details).
}
\details{
...
}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt}

#' @examples
# add example
}
\keyword{print}
