\name{inc_inf-class}


\alias{inc_inf-class}
\alias{summary.inc_inf-class}
\alias{print.inc_inf}
\alias{summary.inc_inf}

\title{
Methods for the class \code{inc_inf}.
}
\description{
Various methods for the output of the function \code{\link{inc_inf}}.
}
\usage{
\method{print}{inc_inf}(x,...)
\method{summary}{inc_inf}(object,...)
}
\arguments{
    \item{x}{An object of class \code{inc_inf}.}
    \item{object}{An object of class \code{inc_inf}.}
        \item{\dots}{arguments to be passed}
}

\seealso{
\code{\link{inc_inf}}
}
