% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_md.R
\name{inc_md}
\alias{inc_md}
\alias{print.inc_md}
\alias{summary.inc_md}
\alias{print.summary.inc_md}
\title{Title line}
\usage{
inc_md(sims, t, t1, t2, t3)

\method{print}{inc_md}(x, ...)

\method{summary}{inc_md}(object, ...)

\method{print}{summary.inc_md}(x, ...)
}
\arguments{
\item{sims}{R object ...}

\item{t}{...}

\item{t1}{...}

\item{t2}{...}

\item{t3}{...}

\item{x}{R object of class \code{inc_md} or  \code{summary.inc_md}.}

\item{\dots}{Additional arguments (ignored)}

\item{object}{R object of class \code{inc_md}.}
}
\value{
Description of returned R object (data frame?)
}
\description{
Brief description (more details could be provided under Details).
}
\details{
...
}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt}

#' @examples
# add example
}
\keyword{print}
