% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_oth.R
\name{inc_oth}
\alias{inc_oth}
\alias{print.inc_oth}
\alias{summary.inc_oth}
\alias{print.summary.inc_oth}
\title{Title line}
\usage{
inc_oth(sims, t, t1, t2, t3, null_effect)

\method{print}{inc_oth}(x, ...)

\method{summary}{inc_oth}(object, ...)

\method{print}{summary.inc_oth}(x, ...)
}
\arguments{
\item{sims}{R object ...}

\item{t}{...}

\item{t1}{...}

\item{t2}{...}

\item{t3}{...}

\item{null_effect}{...}

\item{x}{R object of class \code{inc_oth} or  \code{summary.inc_oth}.}

\item{\dots}{Additional arguments (ignored)}

\item{object}{R object of class \code{inc_oth}.}
}
\value{
Description of returned R object (data frame?)
}
\description{
Brief description (more details could be provided under Details).
}
\details{
...
}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt}

#' @examples
# add example
}
\keyword{print}
