% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_smd.R
\name{inc_smd}
\alias{inc_smd}
\alias{print.inc_smd}
\alias{summary.inc_smd}
\alias{print.summary.inc_smd}
\title{Title line}
\usage{
inc_smd(sims, t1 = 0.2, t2 = 0.5, t3 = 0.8)

\method{print}{inc_smd}(x, ...)

\method{summary}{inc_smd}(object, ...)

\method{print}{summary.inc_smd}(x, ...)
}
\arguments{
\item{sims}{R object ...}

\item{t1}{...}

\item{t2}{...}

\item{t3}{...}

\item{x}{R object of class \code{inc_smd} or  \code{summary.inc_smd}.}

\item{\dots}{Additional arguments (ignored)}

\item{object}{R object of class \code{inc_smd}.}
}
\value{
Description of returned R object (data frame?)
}
\description{
Brief description (more details could be provided under Details).
}
\details{
...
}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt}

#' @examples
# add example
}
\keyword{print}
