\name{inc_subset-class}


\alias{inc_subset-class}
\alias{summary.inc_subset-class}
\alias{print.inc_subset}
\alias{summary.inc_subset}

\title{
Methods for the class \code{inc_subset}.
}
\description{
Various methods for the output of the function \code{\link{inc_subset}}.
}
\usage{
\method{print}{inc_subset}(x,...)
\method{summary}{inc_subset}(object,...)
}
\arguments{
    \item{x}{An object of class \code{inc_subset}.}
    \item{object}{An object of class \code{inc_subset}.}
        \item{\dots}{arguments to be passed}
}

\seealso{
\code{\link{inc_subset}}
}
