\name{inc_t_sens-class}


\alias{inc_t_sens-class}
\alias{print.inc_t_sens}

\title{
Methods for the class \code{inc_t_sens}.
}
\description{
Various methods for the output of the function \code{\link{inc_t_sens}}.
}
\usage{
\method{print}{inc_t_sens}(x,...)
}
\arguments{
    \item{x}{An object of class \code{inc_t_sens}.}
        \item{\dots}{arguments to be passed}
}

\seealso{
\code{\link{inc_t_sens}}
}
