% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf.R
\name{inf}
\alias{inf}
\alias{print.inf}
\alias{summary.inf}
\title{Decision Inconsistency index for individual studies and leave-one out sensitivity analysis for inconsistency measures}
\usage{
inf(x)

\method{print}{inf}(x, ...)

\method{summary}{inf}(object, ...)
}
\arguments{
\item{x}{An R object of class \code{inc}.}

\item{\dots}{Additional arguments (ignored).}

\item{object}{An R object of class \code{inf}.}
}
\value{
The \code{inf} function generates:
\item{di_stud}{A dataframe displaying the Decision Inconsistency Index for each individual primary study.}
\item{di_asi_sens}{A dataframe displaying the results of the leave-one-out sensitivity analysis for the Decision Inconsistency index and the Across-Studies Inconsistency index.}
}
\description{
Computation of the Decision Inconsistency index for each individual primary study and leave-one-out sensitivity analysis for the Decision Inconsistency index and the Across-Studies Inconsistency index.
}
\details{
The \code{inf} function (i) computes the Decision Inconsistency index (DI)
for each individual primary study and (ii) performs leave-one-out
sensitivity analysis for the DI and for the Across-Studies Inconsistency
index. This function takes, as its single argument, an object of the class
\code{inc}.
}
\examples{
data(anticoagulation)
inc_anticoagulation <-
  inc(log(anticoagulation), t1 = 20, t2 = 30, t3 = 40, br = 0.5, sm = "or")
sens <- inf(inc_anticoagulation)
sens

}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{print}
