% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metainc-package.R
\docType{package}
\name{metainc-package}
\alias{metainc-package}
\title{metainc: Brief overview of methods and general hints}
\description{
R package \bold{metainc} ...
}
\details{
R package \bold{metainc} provides the following meta-analysis methods:
\itemize{
\item ... ;
\item ... ;
\item ... .
}

Type \code{help(package = "metainc")} for a listing of R functions
available in \bold{metainc}.

Type \code{citation("metainc")} on how to cite \bold{metainc} in
publications.

The development version of \bold{metainc} is available on GitHub
\url{https://github.com/}.
}
\references{
Add reference(s)
}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{package}
