% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montelukast.R
\docType{data}
\name{montelukast}
\alias{montelukast}
\title{montelukast}
\format{
A matrix with 5000 simulated mean differences (rows) for 9 studies
(columns).
}
\source{
Add reference to source (if available).
}
\description{
Matrix montelukast
}
\examples{
data(montelukast)
inc(montelukast, t1 = 0.2, t2 = 0.4, t3 = 0.6, sm = "md")
}
\seealso{
\code{\link{inc}}
}
\keyword{datasets}
