% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_br.R
\name{sens_br}
\alias{sens_br}
\alias{plot.sens_br}
\title{Sensitivity analysis (based on the baseline risk) for the Decision Inconsistency index and Across-Studies Inconsistency index}
\usage{
sens_br(x, br1, br2, dt1, sm, by = 0.01, scale = 1000)

\method{plot}{sens_br}(
  x,
  ylim1 = c(0, 100),
  ylim2 = c(0, 100),
  ylab1 = "DI index (\%)",
  ylab2 = "ASI index (\%)",
  ...
)
}
\arguments{
\item{x}{An R object of class \code{sims} or a matrix containing the 
simulated effect sizes of primary studies. Note, transformed effect
sizes must be provided, for example, log odds ratios instead of odds
ratios.}

\item{br1}{Smallest baseline risk considered.}

\item{br2}{Largest baseline risk considered.}

\item{dt1}{A single numeric defining the decision threshold to distinguish
meaningful from trivial effects.}

\item{sm}{A character string indicating the summary measure used in
primary studies (either \code{sm = "OR"}, \code{sm = "RR"} or 
\code{sm = "HR"}).}

\item{by}{Increment of the sequence from \code{br1} to \code{br2}.}

\item{scale}{The number of people per which absolute decision thresholds are
provided (default: 1000, i.e., absolute decision threshold values are
defined per 1000 people).}

\item{ylim1}{The y limits (min, max) of the plot showing the Decision
Inconsistency index.}

\item{ylim2}{The y limits (min, max) of the plot showing the Across-Studies
Inconsistency index.}

\item{ylab1}{A label for the y-axis (Decision Inconsistency index).}

\item{ylab2}{A label for the y-axis (Across-Studies Inconsistency index).}

\item{\dots}{Additional graphical arguments (ignored).}
}
\value{
A dataframe containing
\item{br}{Baseline risk}
\item{ASI}{Decision Inconsistency index at baseline risk}
\item{DI}{Across-Studies Inconsistency index at Baseline risk}
}
\description{
Sensitivity analysis on the Decision Inconsistency index and the
Across-Studies Inconsistency index based on a range of baseline risks.
It is applicable only to meta-analyses with binary outcome data (effect size
measures expressed as risk ratios, odds ratios or hazard ratios).
}
\details{
The \code{\link{inc}} function computes the Decision Inconsistency index
(DI) and the Across-Studies Inconsistency index (ASI) for a single
baseline risk. This function allows for performing sensitivity analysis
according to the baseline risk. It can only be applied for meta-analyses
with binary outcome data (effect size measures expressed as risk ratios,
odds ratios or hazard ratios), with the DI and the ASI being calculated
based on absolute effects. As a result, the decision threshold value
(\code{t}) must be provided as an absolute effect (i.e., the minimum number
of additional or diminished events that should occur for the effect to be
considered appreciable / important instead of trivial). By default, it is
assumed that this threshold value is provided per 1000 people (argument
\code{scale = 1000}).
}
\examples{
data(anticoagulation)
dis <- sens_br(log(anticoagulation),  br1 = 0.3, br2 = 0.7, dt1 = 20,
  sm = "OR", by = 0.1)
dis
plot(dis, ylim1 = c(50, 70), ylim2 = c(20, 30))

}
\references{
Schunemann HJ, Higgins JPT, Vist GE, et al. (2019).
\dQuote{Completing ‘Summary of findings’ tables and grading the certainty of
the evidence.}
\emph{Cochrane Handbook for Systematic Reviews of Interventions},
375--402.

Skoetz N, Goldkuhle M, van Dalen EC, et al. (2020).
\dQuote{GRADE guidelines 27: how to calculate absolute effects for
time-to-event outcomes in summary of findings tables and Evidence Profiles.}
\emph{Journal of Clinical Epidemiology}, \bold{118},
124--131.
}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
