% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_dt.R
\name{sens_dt}
\alias{sens_dt}
\alias{heatplot}
\title{Sensitivity analysis (based on decision thresholds) for the Decision Inconsistency index and Across-Studies Inconsistency index}
\usage{
sens_dt(x, br = NULL, min1, max1, min2, max2, sm, by = 1, scale = 1000)

heatplot(x, limits1 = NULL, limits2 = NULL, ...)
}
\arguments{
\item{x}{An R object of class \code{sims} or a matrix containing the 
simulated effect sizes of primary studies. Note, transformed effect
sizes must be provided, for example, log odds ratios instead of odds
ratios.}

\item{br}{Baseline risk (only considered for odds, risk or hazard ratio).}

\item{min1}{A single numeric defining the lowest value for the lower
decision threshold. Must be lower than \code{max1}, \code{min2} and
 \code{max2}.}

\item{max1}{A single numeric defining the highest value for the lower
decision threshold. Must be higher than \code{min1}, but lower than
\code{min2} and \code{max2}.}

\item{min2}{A single numeric defining the lowest value for the higher
decision threshold. Must be higher than \code{min1} and \code{max1},
but lower than \code{max2}.}

\item{max2}{A single numeric defining the highest value for the higher
decision threshold. Must be higher than \code{min1}, \code{max1} and
\code{min2}.}

\item{sm}{A character string indicating the summary measure used in
primary studies (either \code{sm = "OR"}, \code{sm = "RR"} or 
\code{sm = "HR"}).}

\item{by}{Increment of the sequences from \code{min1} to \code{max1} and
\code{min2} to \code{max2}.}

\item{scale}{The number of people per which absolute decision thresholds are
provided (default: 1000, i.e., absolute decision threshold values are
defined per 1000 people).}

\item{limits1}{Limits for the colour range in the heatplot showing the
Decision Inconsistency index.}

\item{limits2}{Limits for the colour range in the heatplot showing the
Across-Studies Inconsistency index.}

\item{\dots}{Additional graphical arguments (ignored).}
}
\value{
A dataframe containing
\item{dt}{Decision threshold}
\item{ASI}{Decision Inconsistency index at baseline risk}
\item{DI}{Across-Studies Inconsistency index at Baseline risk}
}
\description{
Sensitivity analysis on the Decision Inconsistency index and the
Across-Studies Inconsistency index based on a range of decision thresholds.
}
\details{
The \code{\link{inc}} function computes the Decision Inconsistency index
(DI) and the Across-Studies Inconsistency index (ASI) for a single
baseline risk. This function allows for performing sensitivity analysis
according to the decision thresholds.

 Two possibilities are considered:
 \itemize{
   \item The DI and the ASI are calculated based on absolute effects for meta-analyses of odds ratios, risk ratios or hazard ratios: This requires the setting of a baseline risk (i.e., \code{br} must be defined), as well as the definition of the effect size measure computed for the primary studies (i.e., \code{sm} must be defined either as “odds ratio” [\code{"or"}], “risk ratio” [\code{"rr"}] or “hazard ratio” [\code{"hr"}]). The decision threshold values (\code{min1}, \code{max1}, \code{min2}, \code{max2}) should be provided as absolute effects (i.e., minimum number of additional or diminished events that should occur for the effect to be considered appreciable/important instead of trivial). By default, it is assumed that these threshold values are provided per 1000 people (argument \code{scale}).
   \item The DI and the ASI are calculated by directly comparing the effect size measures of primary studies with decision thresholds (\code{min1}, \code{max1}, \code{min2}, \code{max2}). This is the most appropriate approach except when assessing absolute effects for meta-analysis of odds ratios, risk ratios or hazard ratios. The \code{br}, \code{sm} and \code{den} arguments should not be defined.
 }
}
\examples{
data(anticoagulation)
dis <- sens_dt(log(anticoagulation),
  br = 0.5, min1 = -30, max1 = -10, min2 = 10, max2 = 30, sm = "or")
#
head(dis)
summary(dis$DI)
summary(dis$ASI)

heatplot(dis)

}
\references{
Schunemann HJ, Higgins JPT, Vist GE, et al. (2019).
\dQuote{Completing ‘Summary of findings’ tables and grading the certainty of
the evidence.}
\emph{Cochrane Handbook for Systematic Reviews of Interventions},
375--402.

Skoetz N, Goldkuhle M, van Dalen EC, et al. (2020).
\dQuote{GRADE guidelines 27: how to calculate absolute effects for
time-to-event outcomes in summary of findings tables and Evidence Profiles.}
\emph{Journal of Clinical Epidemiology}, \bold{118},
124--131.
}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
