% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anticoagulation.R
\docType{data}
\name{anticoagulation}
\alias{anticoagulation}
\title{Samples of effect sizes (odds ratio)}
\format{
A matrix with 5000 sampled odds ratios (rows) for 18 studies
(columns) following a Bayesian meta-analysis.
}
\source{
Akl, E. A., Kahale, L. A., Hakoum, M. B., Matar, C. F., Sperati, F.,
Barba, M., et al. (2017).
\dQuote{Parenteral anticoagulation in ambulatory patients with cancer.}
\emph{Cochrane Database of Systematic Reviews},
\bold{9}: CD006652.
}
\description{
Matrix with samples of effect sizes from a meta-analysis by Akl et
al. (2017).
}
\details{
The \code{anticoagulation} data is from Akl et al. (2017) and
displays results presented as odds ratios. Each column corresponds
to a different primary study.
}
\examples{
data(anticoagulation)

# Since the results are already presented as odds ratios, we need
# to indicate that effects have not been transformed to log odds
# ratios yet (\code{transf = FALSE}).
dis1 <- inc(anticoagulation, br = 0.504, dt1 = 16, sm = "OR",
  transf = FALSE)
dis1

# Alternatively, we may simply apply the \code{inc()} function to
# log odds ratios.
data(anticoagulation)
dis2 <- inc(log(anticoagulation), br = 0.504, dt1 = 16, sm = "OR")
dis2
}
\seealso{
\code{\link{inc}}
}
\keyword{datasets}
