% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anticoagulation.R
\docType{data}
\name{anticoagulation}
\alias{anticoagulation}
\title{anticoagulation}
\format{
A matrix with 5000 simulated odds ratios (rows) for 18 studies
(columns).
}
\source{
Add reference to source (if available).
}
\description{
Matrix anticoagulation
}
\examples{
data(anticoagulation)
dis <- inc(log(anticoagulation), br = 0.3, t1 = 20, sm = "OR")
dis
}
\seealso{
\code{\link{inc}}
}
\keyword{datasets}
