% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anticoagulation_df.R
\docType{data}
\name{anticoagulation_df}
\alias{anticoagulation_df}
\title{anticoagulation_df}
\format{
A dataset with 17 studies and 7 variables.
}
\source{
Add reference to source (if available).
}
\description{
Meta-analysis dataset anticoagulation_df
}
\examples{
data("anticoagulation_df")
m1 <- meta::metagen(yi, sqrt(vi), sm = "OR", data = anticoagulation_df,
  studlab = LETTERS[1:18])
set.seed(1090) # Make simulated effect sizes reproducible
sims1 <- getsims(m1)
sims1

\dontrun{
# Same simulated effect sizes using R package metafor (must be installed)

m1f <- metafor::rma(anticoagulation_df, measure = "OR", slab = LETTERS[1:18])
set.seed(1090) # Make simulated effect sizes reproducible
sims2 <- getsims(m1f)
sims2

all.equal(sims1, sims2) # Only difference: package name
}
}
\seealso{
\code{\link{inc}}
}
\keyword{datasets}
