% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsims.R
\name{getsims}
\alias{getsims}
\title{Title line}
\usage{
getsims(obj, param, package, nsims = 10000, t_exp = FALSE)
}
\arguments{
\item{obj}{R object ...}

\item{param}{...}

\item{package}{...}

\item{nsims}{...}

\item{t_exp}{A logical ...}
}
\value{
Description of returned R object (data frame?)
}
\description{
Brief description (more details could be provided under Details).
}
\details{
More specific details on R function.
}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt}

#' @examples
# add example
}
