% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.R, R/inf.R
\name{inc}
\alias{inc}
\alias{print.inc}
\alias{summary.inc}
\alias{print.summary.inc}
\alias{print.summary.inf}
\title{Decision Inconsistency and Across-Studies Inconsistency index}
\usage{
inc(
  x,
  t1,
  t2,
  t3,
  sm,
  br = NULL,
  scale = 1000,
  transf = TRUE,
  transf.dt = FALSE
)

\method{print}{inc}(x, digits = 1, ...)

\method{summary}{inc}(object, ...)

\method{print}{summary.inc}(x, digits = 1, ...)

\method{print}{summary.inf}(x, ...)
}
\arguments{
\item{x}{An R object of class \code{sims} or a matrix containing the 
simulated effect sizes of primary studies.}

\item{t1}{A single numeric defining the decision threshold to distinguish
(i) meaningful from trivial effects, if arguments \code{t2} and \code{t3}
are not provided, (ii) negative from trivial effects, if only argument
\code{t2} is also provided or (iii) small from trivial effects if
arguments \code{t2} and \code{t3} are provided.}

\item{t2}{A single numeric defining the decision threshold to distinguish
(i) positive from trivial effects if argument \code{t3} is not provided
or (ii) moderate from small effects, if argument \code{t3} is provided.}

\item{t3}{A single numeric defining the decision threshold to distinguish
large from moderate effects.}

\item{sm}{A character string indicating the summary measure used in
primary studies (see Details).}

\item{br}{Baseline risk (only considered for odds, risk or hazard ratio).}

\item{scale}{The number of people per which absolute decision thresholds are
provided (default: 1000, i.e., absolute decision threshold values are
defined per 1000 people). Only considered if \code{br} is not missing.}

\item{transf}{A logical indicating whether the values of an effect size
matrix (argument \code{x}) are transformed. By default
(\code{transf = TRUE}), it is assumed that the matrix contains log odds
ratios instead of odds ratios, for example.}

\item{transf.dt}{A logical indicating whether relative decision thresholds are
transformed or on the original scale. If \code{transf.dt = FALSE} (default),
relative decision thresholds are expected to be odds ratios instead of
log odds ratios for \code{sm = "OR"}, for example.}

\item{digits}{Minimal number of significant digits to print percentages, see
\code{print.default}.}

\item{\dots}{Additional arguments (ignored)}

\item{object}{R object of class \code{inc}.}
}
\value{
An object of class \code{inc}, for which some standard methods are available, see \code{\link{metainc-package}}. Some of the components include:
\item{DI}{A percentage corresponding to the Decision Inconsistency index. The higher/closer to 100\% the value, the higher the inconsistency.}
\item{ASI}{A percentage corresponding to the Across-Studies Inconsistency index. The higher/closer to 100\% the value, the higher the across-studies inconsistency.}
\item{class_distribution}{A data frame containing the proportion of simulations indicating:
    \itemize{
      \item Large positive effects (effect sizes higher than \code{tl}): “large (higher)” row;
      \item Moderate positive effects (efect sizes between \code{tm} and \code{tl}): “moderate (higher)” row;
      \item Small positive effects (effect sizes between \code{ts} and \code{tm}): “small (higher)” row;
      \item Non meaningful effects (effect sizes between \code{-ts} and \code{ts}): “not meaningful” row;
      \item Small negative effects (effect sizes between \code{-ts} and \code{-tm}): “small (lower)” row;
      \item Moderate negative effects (effect sizes between \code{-tm} and \code{-tl}): “moderate (lower)” row;
      \item Large negative effects (effect sizes lower than \code{-tl}): “large (lower)” row.
    }
}
\item{prop_over_null}{A numeric value indicating the proportion of simulations with a value higher than the value representing no difference between the groups.}
}
\description{
Calculate Decision Inconsistency (DI) and Across-Studies Inconsistency (ASI)
index
}
\details{
This function calculates the Decision Inconsistency index (DI) and the
Across-Studies Inconsistency index (ASI) for meta-analyses. The following
possibilities are considered depending on the type of effect size measures:
  \itemize{
    \item Effect size measure corresponding to a ratio (\code{sm = "OR"}, \code{"RR"} or \code{"HR"}) with the DI and the ASI being calculated based on absolute effects: This requires the setting of a baseline risk (i.e., \code{br} must be defined). The decision threshold values (\code{t1}, \code{t2} and \code{t3} must be provided as absolute effects (i.e., number of additional or diminished events per N people. By default, it is assumed that these threshold values are provided per 1000 people. However, this can be changed using the \code{scale} argument).
    \item Effect size measure corresponding to a ratio (\code{sm} = \code{"OR"}, \code{"RR"}, \code{"HR"} or \code{"GEN_ratio"}) with the DI and the ASI being calculated based on relative effect size measures: The simulation results of relative effect size measures of primary studies are directly compared with decision thresholds (\code{t1}, \code{t2}, \code{t3}) also expressed as relative effect size measures. This is the adopted approach when no information is provided on the baseline risk (\code{br}).
    \item Effect size measure corresponding to a difference (\code{sm = "MD"}, \code{"SMD"}, \code{"RD"} or \code{"GEN_diff"}): The simulation results of the effect size measures of primary studies are directly compared with decision thresholds (\code{t1}, \code{t2}, \code{t3}) also expressed as differences.
  }
  
Of note, when dealing with relative effect size measures, judgements based
on absolute effects tend to be considered more important for
decision making. The formulae for calculating absolute effects based on
relative effect size measures are those used in the GRADE approach
(see References below).

Either only argument \code{t1} or arguments  \code{t1}, \code{t2} and
\code{t3} must be provided.

Argument \code{sm} must be \code{"OR"} (odds ratio),
\code{"RR"} (risk ratio), \code{"HR"} (hazard ratio),
\code{"MD"} (mean difference), \code{"SMD"} (standardised mean difference),
\code{"RD"} (risk difference), \code{"GEN_diff"} (generic difference), or
\code{"GEN_ratio"} (generic ratio).

The baseline risk (optional argument \code{br} which must be a numeric value
between 0 and 1) can be provided when \code{sm = "OR"}, \code{"RR"} or
\code{"HR"}. The baseline risk is also known as assumed comparator risk,
i.e., the risk that the outcome of interest occurs in the comparison
intervention.
}
\examples{
# Example with effect sizes measures expressed as ratios and with
# calculation of the Decision Inconsistency index and the Across-Studies
# Inconsistency index being calculated based on absolute effects:

data(anticoagulation)
inc_anticoagulation <-
  inc(anticoagulation, t1 = 20, t2 = 30, t3 = 40, br = 0.5, sm = "or",
      transf = FALSE)
inc_anticoagulation

\dontrun{
# Same result
inc_anticoagulation <-
  inc(log(anticoagulation), t1 = 20, t2 = 30, t3 = 40, br = 0.5, sm = "or")
inc_anticoagulation

# Example with calculation of the Decision Inconsistency index and the 
# Across-Studies Inconsistency index based on effect size measures expressed
# as mean differences:

data(montelukast)
inc_montelukast <- inc(montelukast, t1 = 0.2, t2 = 0.4, t3 = 0.6, sm = "md")
inc_montelukast
}

}
\references{
Cohen J. (1998).
\dQuote{Statistical Power Analysis in the Behavioral Sciences},
2nd edition ed. Hillsdale (NJ): Lawrence Erlbaum Associates, Inc.

Schunemann HJ, Higgins JPT, Vist GE, et al. (2019).
\dQuote{Completing ‘Summary of findings’ tables and grading the certainty of
the evidence.}
\emph{Cochrane Handbook for Systematic Reviews of Interventions},
375--402.

Schunemann HJ, Vist GE, Higgins JPT, et al. (2019).
\dQuote{Interpreting results and drawing conclusions.}
\emph{Cochrane Handbook for Systematic Reviews of Interventions},
403--431.

Skoetz N, Goldkuhle M, van Dalen EC, et al. (2020).
\dQuote{GRADE guidelines 27: how to calculate absolute effects for
time-to-event outcomes in summary of findings tables and Evidence Profiles.}
\emph{Journal of Clinical Epidemiology}, \bold{118},
124--131.
}
\author{
Bernardo Sousa-Pinto \email{bernardo@med.up.pt},
  Guido Schwarzer \email{guido.schwarzer@uniklinik-freiburg.de}
}
\keyword{print}
